When a test fails, run it first the next time
This patch makes run_job() return a very large "elapsed time" for jobs
that fail. The effect is that on the next run, tests that previously
failed will be tried first regardless of their actual runtime; since
tests that failed before are likely to fail again, and since observing
failures is often the reason one runs tests in the first place, this
will usually greatly reduce the time one has to wait.