Always print a newline before exiting
We used to not print a newline at the end if the last line was a
transient. It looked OK anyway, because transient lines are suffixed
with just enough spaces to fill the line, but it messed with at least
one user's colored bash prompt. And ending with a newline is the right
thing to do anyway.