Log to gtest-parallel-logs/ subdir.
This changes the default logdir from /tmp/gtest-parallel to
/tmp/gtest-parallel-logs. The primary motivation is that that we do this by
appending gtest-parallel-logs (to /tmp by default), so that if a user provides a
directory that is already in use, we wouldn't clear that directory out
completely (--output_dir=Docs/ was previously dangerous, as we'd nuke Docs/*).

Now instead we'd only clear out a subdirectory that we create and own ourselves.

Fixes #44.