log_exit(): Add handler for signals
On 'nix systems the return code might be negative indicating a signal
number. Rather than simply printing a negative return code, try to
interpret the signal and print the signal name for behavior that is
closer to what a user would see if they ran the test outside of
gtest-parallel.

Fixes: #98